[Step4] 로또(수동) (#944)
* docs: step4 요구사항을 반영

* feat: 수동으로 번호를 입력받은 경우 처리 가능한 도메인 로직 작성

* docs: 수동으로 구매할 로또 수를 입력받는 기능에 대한 정의 작성

* feat: inputview를 통해 수동으로 로또 번호를 입력받는 기능 작성

* refactor: Collection이 empty인지 판별하는 유틸성 메서드 분리

* feat: 수동 자동으로 구매한 로또 번호 수를 출력하는 기능 추가

* feat: 예외 발생시 메인에서 출력할 수 있도록 하는 기능 추가

* docs: doc 오타 수정

* Revert "feat: 예외 발생시 메인에서 출력할 수 있도록 하는 기능 추가"

This reverts commit ad618c821274999f9dd69eb980c359d862fe0ee7.

* bugfix: 1등 2번 당첨시 오버플로우 발생 문제 픽스

* bugfix: 0인 경우 로또 번호가 생성되는 버그 수정

* bugfix: 수동으로 구매하기 원하는 로또 갯수가 0 혹은 음수일 경우 0으로 변경 후 수동 입력을 받지 않도록 변경

* bugfix: 수동으로 구매하기 원하는 로또가 지불한 금액보다 큰 경우 예외 반환

* refactor: final 키워드 추가

* refactor: slip 대신 manual 단어 사용

* bugfix: 보너스번호와 당첨번호 중복 경우 처리

* refactor: 상수 이용

* refactor: lotto number 캐싱

* refactor: rank계산 helper 클래스를 WinningLotto 클래스로 마이그레이션

+++ src/main/java/lotto/domain/Money.java
@@ -8,18 +8,18 @@ public class Money {
     public static final String MONEY_MUST_NOT_BE_NEGATIVE = "money must not be negative";
     public static final String MONEY_MUST_BE_DIVIDED = "money must be divided";
     public static final Money ZERO = of(0);
-    private final int value;
+    private final long value;
 
-    private Money(final int money) {
+    private Money(final long money) {
         this.value = money;
     }
 
-    public static Money of(int money) {
+    public static Money of(long money) {
         checkArgument(money >= 0, MONEY_MUST_NOT_BE_NEGATIVE);
         return new Money(money);
     }
 
-    public int divideWithoutRemainder(final Money money) {
+    public long divideWithoutRemainder(final Money money) {
         checkArgument(isDivided(money), MONEY_MUST_BE_DIVIDED);
         return value / money.value;
     }
@@ -51,7 +51,7 @@ public class Money {
         return value == 0;
     }
 
-    public int getValue() {
+    public long getValue() {
         return value;
     }
